Move curry to utils
